+++ src/spatial/index.ts
@@ -1,3 +1,4 @@
 export * from './spatialClient';
+export * from './snapping';
 export * from './types';
 export * from './viewport';